[feat] 전체문화정보화면 정렬 및 페이지네이션(CUL-272)

+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/performance/repository/PerformanceLikeRepository.java
@@ -31,5 +31,5 @@ public interface PerformanceLikeRepository extends JpaRepository<PerformanceLike
     FROM PerformanceLike pl
     WHERE pl.performance.id = :performanceId
 """)
-    int countByPerformanceId(@Param("performanceId") Long performanceId);
+    Long countByPerformanceId(@Param("performanceId") Long performanceId);
 }
